requirements.txt
fastapi
uvicorn[standard]
httpx
feedparser
beautifulsoup4
trafilatura
python-slugify
